requirements.txt
# Python 2.7
azure-batch==5.0.0
python-dotenv==0.8.0